fix : 백엔드 오류 수정 완료

+++ backend/src/main/java/com/tripai/backend/repository/FavoriteEventMapper.java
@@ -12,7 +12,7 @@ public interface FavoriteEventMapper {
     public List<EventSummary> findEventsByUserId(  @Param("userId") Long userId, 
                                             @Param("offset") Integer offset, 
                                             @Param("size") Integer size);
-    public Integer countEventsByUserId(@Param("userId") Long userId);
+    public long countEventsByUserId(@Param("userId") Long userId);
     public Integer deleteEventByEventId(   @Param("userId") Long userId,
                                     @Param("eventId") String eventId);
 }
